requirements.txt
# Gmail API dependencies
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# LLM API clients
openai>=1.0.0
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
